updating gruop controller to accomodate poker game

+++ src/actions/group/WagerSelector.java
@@ -3,6 +3,6 @@ package actions.group;
 @FunctionalInterface
 public interface WagerSelector {
 
-    int getBet(int min, int max);
+    double getBet(double min, double max);
 
 }
